pyproject.toml
[tool.poetry]
name = "company-structure-api"
version = "0.1.0"
description = ""
authors = ["drforse <george.lifeslice@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
django-filter = "^24.3"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
